mbedtls: Move local mbedtls to v3.6.0
The in-tree mbedtls (used for the simulator and some targets) is a few years
old, and currently is unable to pass the rsa tests when built with clang.
Update this mbed TLS to the v3.6.0 release.  This fixes clang support in the
simulator.

There are a few minor changes to configuration and what files are needed to
support newer version of Mbed TLS.

Fixes #1986

Signed-off-by: David Brown <david.brown@linaro.org>
(cherry picked from commit fa4ca87a7b820bff8a6e26a608d6ef12cbc5b819)