boot: zephyr: fix duplicate symbols with tinycrypt
When building Zephyr with mcuboot, with tinycrypt instead of
mbedtls, there are duplicate symbols between the tinycrypt
module and the in-tree copy of tinycrypt that is included with
mcuboot.

The issues are only with symbols in the file

${TINYCRYPT_DIR}/source/utils.c

If Zephyr is being built with the tinycrypt module present,
prefer that over the in-tree copy.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>