zephyr: Fix TLV area was included in trailer size when rounding up
In order to determine the maximum image size, the size of the trailer is
computed. When using swap-move or swap-offset, this trailer size has to
be rounded up to the next multiple of the sector size. However, the
current logic was rouding up the sum of the trailer size and the TLV
area size, instead of only the trailer size. This commit fixes the
issue.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>
(cherry picked from commit f9e4e529a12d0b890a4665e6756aa2c9157aae7d)